lab4/requirements.txt
python-dotenv==1.0.0
redis==5.0.1
kafka-python==kafka-python-ng